1.0.0-1BX11 - Fixed glue/handlers/handle_vsite.pl to only run alias edits on destroy if there are lists for that site. - Added missing 'use Sauce::Util' reference.

+++ BlueOnyx/5106R/ui/base-mailman.mod/glue/handlers/handle_vsite.pl
@@ -1,11 +1,12 @@
-#!/usr/bin/perl -w -I/usr/sausalito/perl -I.
-# $Id: handle_vsite.pl,v 1.0.0-3 Tue 26 Apr 2011 12:12:20 AM CEST
+#!/usr/bin/perl -I/usr/sausalito/perl -I.
+# $Id: handle_vsite.pl,v 1.0.0-4 Sun 01 May 2011 01:49:36 AM CEST
 # Copyright 2011 Team BlueOnyx. All rights reserved.
 #
 # handles the creation of a virtual site's mailman infrastructure
 
 use lib qw( /usr/sausalito/perl );
 use CCE;
+use Sauce::Util;
 
 my $DEBUG = 0;
 $DEBUG && warn `date`." $0\n";
@@ -28,46 +29,45 @@ my $oid = $cce->event_oid();
 # this runs in the CONFIGURE stage which is the same stage as the handler
 # that sets the Vsite.name property, so make sure on create the name
 # property is defined before proceeding
-if ($cce->event_is_create() && ($obj->{name} eq ''))
-{
+if ($cce->event_is_create() && ($obj->{name} eq '')) {
 	$cce->bye('DEFER');
 	exit(0);
 }
 
-if($cce->event_is_destroy())
-{
-	foreach my $oid (
-		$cce->find("MailList", {'site' => $old->{name}}) )
-	{
-		$cce->destroy($oid); # ignore return on destructive ops
-	}
+if($cce->event_is_destroy()) {
 
-	# Destroy per-site mailman aliases
-	foreach my $alias (@mailman_aliases)
-	{
+	@oids = $cce->find('MailList', {'site' => $old->{name}});
+	if ($#oids < 0) {
+	    # Object not found in CCE
+	}
+	else {
+	    foreach $oid (@oids) {
+		$cce->destroy($oid); # ignore return on destructive ops
+	    }
+
+	    # Destroy per-site mailman aliases
+	    foreach my $alias (@mailman_aliases) {
 		my(@marked_for_death) = $cce->find('ProtectedEmailAlias', {
 			'alias'=>$alias,
 			'fqdn'=>$old->{fqdn},
 			});
-		foreach my $oid (@marked_for_death)
-		{
+		foreach my $oid (@marked_for_death) {
 			my ($ret) = $cce->destroy($oid);
 		}
-	}
+	    }
 
-	# Destroy aliases.mailman entries for mailman aliases
-	my($ok) = Sauce::Util::editfile(
+	    # Destroy aliases.mailman entries for mailman aliases
+	    my($ok) = Sauce::Util::editfile(
 		$aliasfile,
 		*edit_mailmanalias,
 		0,
 		$old->{name},
 		\@mailman_aliases
 		);
-	system('/bin/rm -f /etc/mail/aliases.mailman.backup.*');
-
+	    system('/bin/rm -f /etc/mail/aliases.mailman.backup.*');
+	} # from else
 }
-elsif($new->{fqdn})
-{
+elsif($new->{fqdn}) {
 		# Get the name of the list we need to rename
 		my (@oidsx) = $cce->find("MailList", {'site' => $old->{name}});
 		my ($okx, $objx) = $cce->get($oidsx[0]);
@@ -83,10 +83,8 @@ elsif($new->{fqdn})
 
 }
 
-if($cce->event_is_create())
-{
-	foreach my $alias (@mailman_aliases)
-	{
+if($cce->event_is_create()) {
+	foreach my $alias (@mailman_aliases) {
 		my $action = 
 		my(@oids) = $cce->find('ProtectedEmailAlias', {
 			'alias'=>$alias,
@@ -101,8 +99,7 @@ if($cce->event_is_create())
 	}
 }
 
-if($errors)
-{
+if($errors) {
 	$cce->bye('FAIL', $errors);
 	exit 1;
 }
@@ -112,13 +109,11 @@ exit 0;
 
 
 
-sub edit_mailmanalias
-{
+sub edit_mailmanalias {
 	my($in, $out, $enable, $site, $aliasref) = @_;
 	$DEBUG && warn "edit_mailmanalias invoked: ".join(@_)."\n";
 	my ($config, %public_alii);
-	foreach my $alias (@{$aliasref})
-	{
+	foreach my $alias (@{$aliasref}) {
 		$public_alii{$site.'-'.$alias} = 1;
 
 		if($alias =~ /owner/i) {
@@ -127,12 +122,9 @@ sub edit_mailmanalias
 	}
 	$DEBUG && warn "aliases config:\n$config";
 
-	while(<$in>)
-	{
-		if(/^(\S+):\s/)
-		{
-			if ($public_alii{$1}) 
-			{
+	while(<$in>) {
+		if(/^(\S+):\s/) {
+			if ($public_alii{$1}) {
 				$DEBUG && warn "Skipping: $_";
 				next;
 			}
@@ -143,12 +135,10 @@ sub edit_mailmanalias
 	return 1;
 }
 
-sub edit_list_members
-{
+sub edit_list_members {
 	my ($in, $out, $old_fqdn, $new_fqdn) = @_;
 
-	while (<$in>)
-	{
+	while (<$in>) {
 		s/^([^\@]+)\@$old_fqdn$/$1\@$new_fqdn/;
 		print $out $_;
 	}
